+++ lib/core/api/adapters/subsonic_adapter.dart
@@ -65,9 +65,6 @@ class SubsonicAdapter extends SubsonicProtocolAdapter {
   AdapterCapabilities get capabilities => const AdapterCapabilities(
     ratings: true,
     similarSongs: true,
-    likedSongs: true,
-    download: true,
-    lyrics: true,
     artistBio: true,
     transcoding: true,
     scrobbling: true,